requirements.txt
intersystems-irispython
